Separate support tool pages into individual files
